Update index.html
Updated code to include a scoring system and last commit message was a a mistake i meant to say that it works in the browsers console not in terminal which is different.  If The Odin Project tells me to revisit this then I will add html elements to display results in the DOM.   If not then this project is complete.